Build 9_03 final
It has been decided to rewrite the parts of the Haskell side of the rendering engine that interact with the graphics card using OpenGL (as opposed to OpenGLRaw as currently).  This is in case the visual glitch under investigation is caused by an error in memory management when using pointers to interact with the OpenGLRaw bindings.  This will be done in a branch initially.