Game speed can now be set at the command line
The speedScaling engine parameter can now be set through a command line option, which causes player
movement speed and GPLC virtual machine speed to be scaled.  The motivation is to allow a higher
than standard game speed to be enforced when a player chooses the hardest "Health and Safety
Nightmare" difficulty setting.